Implemented getWritableDatabase() function, it returns a database with write permissions.
Implemented consegueInserirDoencas() function, it tries to insert a disease in the database by calling insereDoenca() function.

Implemented insereDoenca() function, it inserts a disease in the database and checks if its ID is != than the default value.

Implemented consegueInserirVacina() function, it tries to insert a vaccine in the database by calling insereVacina() function.

Implemented insereVacina() function, it inserts a vaccine in the database and checks if its ID is != than the default value.

Implemented consegueLerDoencas() test.
It requests a writable database, creates testing diseases and inserts them on the database. Then, it creates a cursor to retrieve data from the database.
Finally it checks if the data inserted and the data retrieved from the cursor are the same.

Implemented consegueLerVacinas() test.
It requests a writable database, creates testing vaccines and inserts them on the database. Then, it creates a cursor to retrieve data from the database.
Finally it checks if the data inserted and the data retrieved from the cursor are the same.